Refactor(Lsp): Moved handlers to config.
