+++ lua/ds_omega/config/Lsp/core/handlers/filter_diagnostics_by_severity/init.lua
@@ -1,9 +1,9 @@
 ---   Setup commands related to filtering diagnostics
 ---@return keymappings (table) Related keymappings compatible with which_key.
 local function setup_filter_diagnostics_by_severity()
-  require('ds_omega.layers.Lsp.handlers.filter_diagnostics_by_severity.commands')
+  require('ds_omega.config.Lsp.core.handlers.filter_diagnostics_by_severity.commands')
 
-  return require('ds_omega.layers.Lsp.handlers.filter_diagnostics_by_severity.keymappings')
+  return require('ds_omega.config.Lsp.core.handlers.filter_diagnostics_by_severity.keymappings')
 end
 
 return setup_filter_diagnostics_by_severity
