+++ lua/ds_omega/config/Lsp/core/handlers/setup_symbol_highlight_under_cursor.lua
