+++ lua/ds_omega/config/Lsp/core/server_configurations/default.lua
@@ -1,4 +1,4 @@
-local lsp_handlers = require('ds_omega.layers.Lsp.handlers')
+local lsp_handlers = require('ds_omega.config.Lsp.core.handlers')
 
 local default_server_config = {
   on_attach = lsp_handlers.on_attach,
