+++ lua/ds_omega/config/Lsp/core/server_configurations/nil_ls.lua
