+++ lua/ds_omega/config/Lsp/core/server_configurations/texlab.lua
