+++ lua/ds_omega/config/Lsp/core/server_configurations/tsserver.lua
@@ -1,4 +1,4 @@
-local lsp_handlers = require('ds_omega.layers.Lsp.handlers')
+local lsp_handlers = require('ds_omega.config.Lsp.core.handlers')
 local prequire = require('ds_omega.utils').prequire
 
 local function setup_ts_utils(client)
